Serverless!
The backend is now run in Lambda instead of Docker. The frontend deploys to an S3 bucket.

So this no longer requires us to run our own server at all.

It also modernizes React/Redux and switches to reducers instead of our old custom dispatcher/store.

It removes the favorites and line toggle features. We'll want to add those back, but they were incompatible with the reducer model.